Organized view package into subfolders
