+++ app/src/main/AndroidManifest.xml
@@ -14,12 +14,12 @@
         android:theme="@style/AppTheme">
 
         <activity android:name=".view.MainActivity" />
-        <activity android:name=".view.DebtActivity" />
-        <activity android:name=".view.ContactsActivity" />
-        <activity android:name=".view.DetailedGroupActivity" />
-        <activity android:name=".view.GroupCreationActivity" />
+        <activity android:name=".view.debt.DebtActivity" />
+        <activity android:name=".view.contact.ContactsActivity" />
+        <activity android:name=".view.group.DetailedGroupActivity" />
+        <activity android:name=".view.group.GroupCreationActivity" />
         <activity
-            android:name=".view.LoginActivity"
+            android:name=".view.login.LoginActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
