+++ app/src/main/java/com/goayo/debtify/view/group/GroupCreationActivity.java
@@ -1,4 +1,4 @@
-package com.goayo.debtify.view;
+package com.goayo.debtify.view.group;
 
 import android.os.Bundle;
