+++ app/src/main/java/com/goayo/debtify/view/util/PickUsersFragment.java
@@ -1,4 +1,4 @@
-package com.goayo.debtify.view;
+package com.goayo.debtify.view.util;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
