+++ app/src/main/res/layout/activity_debt.xml
@@ -9,7 +9,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.DebtActivity">
+        tools:context=".view.debt.DebtActivity">
 
         <fragment
             android:id="@+id/debtNavHostFragment"
