+++ app/src/main/res/layout/activity_group_creation.xml
@@ -9,7 +9,7 @@
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.GroupCreationActivity">
+        tools:context=".view.group.GroupCreationActivity">
         <!-- The NavHostFragment within the activity_group_creation layout -->
         <fragment
             android:id="@+id/groupCreation_nav_host"
