+++ app/src/main/res/navigation/group_navigation_graph.xml
@@ -8,7 +8,7 @@
     <fragment
         android:id="@+id/groupFragment"
         tools:layout="@layout/group_fragment"
-        android:name="com.goayo.debtify.view.GroupFragment"
+        android:name="com.goayo.debtify.view.group.GroupFragment"
         android:label="GroupFragment" >
         <action
             android:id="@+id/action_groupFragment_to_pickUsersFragment"
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/pickUsersFragment"
-        android:name="com.goayo.debtify.view.PickUsersFragment"
+        android:name="com.goayo.debtify.view.util.PickUsersFragment"
         android:label="PickUsersFragment"
         tools:layout="@layout/pick_users_fragment">
         <action
@@ -28,7 +28,7 @@
     </fragment>
     <fragment
         android:id="@+id/showMembersFragment"
-        android:name="com.goayo.debtify.view.ShowMembersFragment"
+        android:name="com.goayo.debtify.view.group.ShowMembersFragment"
         android:label="ShowMembersFragment"
         tools:layout="@layout/show_members_fragment">
         <action
