+++ app/src/main/res/navigation/settle_debt_navigation_graph.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/settleDebtFragment"
-        android:name="com.goayo.debtify.view.SettleDebtFragment"
+        android:name="com.goayo.debtify.view.debt.SettleDebtFragment"
         android:label="SettleDebtFragment"
         tools:layout="@layout/settle_debt_fragment">
         <action
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/pickUsersFragment"
-        android:name="com.goayo.debtify.view.PickUsersFragment"
+        android:name="com.goayo.debtify.view.util.PickUsersFragment"
         android:label="PickUsersFragment"
         tools:layout="@layout/pick_users_fragment">
         <action
